管理操作周りを修正
