+++ src/domain/model/鉢/管理操作/成長を記録.ts
@@ -1,10 +1,11 @@
 import { UserId } from '@frontend/domain/model/user';
-import { 履歴, 鉢サイズ } from '@frontend/domain/model/履歴';
-import dayjs, { Dayjs } from 'dayjs';
+import { 履歴 } from '@frontend/domain/model/履歴';
+import dayjs from 'dayjs';
 import { StorageRepository } from '@frontend/domain/repository/storage';
 import { FSAppRepository } from '@frontend/domain/repository/firestore';
 import { 鉢 } from '@frontend/domain/model/鉢';
 import { 小画像の生成 } from '@frontend/domain/model/鉢/管理操作/新規作成';
+import { _履歴を適用 } from 'src/domain/model/鉢/管理操作/common';
 
 export namespace _成長を記録する {
   export type Params = {
@@ -55,7 +56,7 @@ export const _成長を記録する = async (params: _成長を記録する.Para
   });
 
   console.log('3. 鉢の情報を更新する');
-  const 更新後の鉢 = item.履歴を適用(植替え履歴, small画像のURL, 画像を更新する);
+  const 更新後の鉢 = _履歴を適用(item, 植替え履歴, small画像のURL, 画像を更新する);
   await FSAppRepository.鉢.snapshotを更新(鉢Id, 更新後の鉢.snapshot, date);
   鉢.events.管理.next({ type: '成長の記録' });
 };
